Refactored and renamed my list component to be re-usable for both the related list and the outfit list. The product data is now passed through a context provider instead.
